fix: Ensure the debug flag is not forwarded (#43)
If the PHPUnit test is executed via a Makefile command with the `--debug` flag, for example for this library:

```
make phpunit --debug
```

Then the `MAKEFLAGS` environment variable is updated and forwarded to
the processes in which the test make commands are executed. This result
in a different output which can easily break tests.