chore(deps): Update and lock PHP-CS-Fixer to PHP 8.1
PHP-CS-Fixer should result in code working with the lowest supported
version of the library which is currently PHP 8.1.

This PR adds a comment in the CI file (adding a test for it would be too
tedious) and enforces it on Composer side via the `#platform` property.
This ensures one can update PHP-CS-Fixer locally even if not executing
PHP 8.1.