feat: Bump min PHP version supported to PHP 8.3 (#63)
There is no strong need to do so for the source code itself. However
this package still uses various tools for which it gets more annoying
the more PHP versions are supported.

I don't have a strong opinion as to which minimal version should be
used, so I decided to align this package with PHPUnit.